move remoteWorkMode related code in dedicated module
